Refactor: split device ID creation and device token verification into separate routes
