+++ 1-src/0-assets/field-sync/input-config-sync/notification-field-config.mts
@@ -1,0 +1,8 @@
+import InputField, { DeviceOSEnum, InputSelectionField, InputType } from "./inputField.mjs";
+
+export const NOTIFICATION_DEVICE_FIELDS:InputField[] = [
+    new InputField({title: 'Device Name', field: 'deviceName', type:InputType.TEXT, validationRegex: new RegExp(/^[a-zA-Z0-9' _-]{1,100}$/), validationMessage: 'Required, 5-100 chars, letters, numbers, dashes, underscores.' }),
+    new InputField({title: 'Token', field: 'deviceToken', type:InputType.TEXT, validationRegex: new RegExp(/^[a-zA-Z0-9.:_-]{1,255}$/), validationMessage: 'Required, 5-100 chars, letters, numbers, dashes, underscores.' }),
+    new InputSelectionField({title: 'Platform', field: 'deviceOS', type: InputType.SELECT_LIST, required: true, selectOptionList: Object.values(DeviceOSEnum)}),
+    new InputField({title: 'Endpoint', field: 'endpointARN', type:InputType.TEXT, validationRegex: new RegExp(/^[a-zA-Z0-9:./=_-]{0,255}$/), validationMessage: 'Required, 5-100 chars, letters, numbers, dashes, underscores.' }),
+];
