+++ 1-src/1-api/2-auth/auth-types.mts
@@ -20,7 +20,6 @@ export interface SubscribePost extends Request {
     }
 }
 
-
 export interface JwtData extends JwtPayload {
     jwtUserID: number;
     jwtUserRole:RoleEnum;
