+++ 1-src/1-api/4-circle/circle.mts
@@ -16,8 +16,8 @@ import { clearImage, clearImageCombinations, uploadImage } from '../../2-service
 import { CircleAnnouncementCreateRequest, CircleImageRequest, JwtCircleClientRequest } from './circle-types.mjs';
 import getCircleEventSampleList from './circle-event-samples.mjs';
 import { ProfileListItem } from '../../0-assets/field-sync/api-type-sync/profile-types.mjs';
-import { sendNotificationSingleRecipient } from '../3-profile/profile-utilities.mjs';
-import { SingleRecipientNotificationType } from '../3-profile/profile-types.mjs';
+import { CircleNotificationType } from '../8-notification/notification-types.mjs';
+import { sendNotificationCircle} from '../8-notification/notification-utilities.mjs';
 
 
 /******************
@@ -318,7 +318,7 @@ export const POST_circleLeaderMemberInvite =  async(request: JwtCircleClientRequ
         const circleItem:CircleListItem = (await DB_SELECT_CIRCLE(request.circleID)).toListItem();
         circleItem.status = CircleStatusEnum.INVITE;
 
-        sendNotificationSingleRecipient(request.jwtUserID, parseInt(request.params.client), SingleRecipientNotificationType.CIRCLE_INVITE, undefined, request.circleID);
+        sendNotificationCircle(request.jwtUserID, [parseInt(request.params.client)], circleItem.circleID, CircleNotificationType.CIRCLE_INVITE);
         response.status(202).send(circleItem);
     }
 };
@@ -347,6 +347,9 @@ export const POST_circleMemberJoinAdmin =  async(request: JwtCircleClientRequest
         await DB_DELETE_CONTACT_CACHE_CIRCLE_MEMBERS(request.circleID);
         const circle:CircleListItem = (await DB_SELECT_CIRCLE(request.circleID)).toListItem();
         circle.status = CircleStatusEnum.MEMBER;
+
+        sendNotificationCircle(request.jwtUserID, [request.clientID], circle.circleID, CircleNotificationType.CIRCLE_INVITE);
+        
         response.status(202).send(circle);
         log.event(`Admin assigning user ${request.clientID} to circle ${request.circleID}`);
     }
