+++ 1-src/1-api/6-partner/partner-request.mts
@@ -8,9 +8,8 @@ import { DATABASE_PARTNER_STATUS_ENUM, DATABASE_USER_ROLE_ENUM } from '../../2-s
 import USER from '../../2-services/1-models/userModel.mjs';
 import { DB_DELETE_CONTACT_CACHE, DB_DELETE_CONTACT_CACHE_BATCH, DB_IS_USER_ROLE, DB_SELECT_USER, DB_SELECT_USER_ROLES } from '../../2-services/2-database/queries/user-queries.mjs';
 import { PartnerListItem, ProfileListItem } from '../../0-assets/field-sync/api-type-sync/profile-types.mjs';
-import { sendNotificationSingleRecipient } from '../3-profile/profile-utilities.mjs';
-import { SingleRecipientNotificationType } from '../3-profile/profile-types.mjs';
-
+import { sendNotification } from '../8-notification/notification-utilities.mjs';
+import { NotificationType } from '../8-notification/notification-types.mjs';
 
 
 /*********************************
@@ -91,8 +90,8 @@ export const POST_PartnerContractAccept = async(request:JwtClientRequest, respon
 
         log.event(`User ${request.jwtUserID} and partner ${request.clientID} are now ${newStatus}`);
 
-        if (newStatus === PartnerStatusEnum.PENDING_CONTRACT_PARTNER) sendNotificationSingleRecipient(getUserID(request.jwtUserID, request.clientID), getPartnerID(request.jwtUserID , request.clientID), SingleRecipientNotificationType.PARTNERSHIP_REQUEST);
-        else if (newStatus === PartnerStatusEnum.PARTNER) sendNotificationSingleRecipient(getPartnerID(request.jwtUserID, request.clientID), getUserID(request.jwtUserID, request.clientID), SingleRecipientNotificationType.PARTNERSHIP_ACCEPT);
+        if (newStatus === PartnerStatusEnum.PENDING_CONTRACT_PARTNER) sendNotification(getUserID(request.jwtUserID, request.clientID), [getPartnerID(request.jwtUserID , request.clientID)], NotificationType.PARTNERSHIP_REQUEST);
+        else if (newStatus === PartnerStatusEnum.PARTNER) sendNotification(getPartnerID(request.jwtUserID, request.clientID), [getUserID(request.jwtUserID, request.clientID)], NotificationType.PARTNERSHIP_ACCEPT);
 
         response.status(200).send(await DB_SELECT_PARTNERSHIP(request.jwtUserID, request.clientID));
     } else
@@ -135,6 +134,10 @@ export const POST_PartnerStatusAdmin = async(status:PartnerStatusEnum|undefined,
 
     if(await DB_ASSIGN_PARTNER_STATUS(request.clientID, request.partnerID, DATABASE_PARTNER_STATUS_ENUM[status])) {
         await DB_DELETE_CONTACT_CACHE_BATCH([request.clientID, request.partnerID]);
+
+        if (status === PartnerStatusEnum.PENDING_CONTRACT_PARTNER) sendNotification(getUserID(request.jwtUserID, request.clientID), [getPartnerID(request.jwtUserID , request.clientID)], NotificationType.PARTNERSHIP_REQUEST);
+        else if (status === PartnerStatusEnum.PARTNER) sendNotification(getPartnerID(request.jwtUserID, request.clientID), [getUserID(request.jwtUserID, request.clientID)], NotificationType.PARTNERSHIP_ACCEPT);
+
         log.event(`ADMIN: User ${request.clientID} and partner ${request.partnerID} are now ${status}`);
         response.status(200).send(await DB_SELECT_PARTNERSHIP(request.clientID, request.partnerID));
     } else
